Split credential storage and harden replacement
