+++ src/delta_exchange_mcp/auth/store.py
@@ -1,60 +1,50 @@
-"""Versioned credentials backed by an operating-system credential store."""
+"""Versioned Delta credentials and their replacement transactions."""
 
-import hashlib
-import io
-import json
 import logging
 import os
-import re
-import secrets
-import stat
-import tempfile
 import threading
-import time
-from collections.abc import Callable, Iterator, Mapping
-from contextlib import AbstractContextManager, contextmanager
-from dataclasses import asdict, dataclass, field, replace as replace_fields
+from collections.abc import Callable, Mapping
+from dataclasses import dataclass, field, replace as replace_fields
 from datetime import UTC, datetime
 from enum import StrEnum
 from pathlib import Path
-from typing import Protocol
 
-import keyring
-from dotenv import dotenv_values
-from dotenv.parser import parse_stream
-from keyring.backend import KeyringBackend
-from keyring.errors import PasswordDeleteError
-
-SERVICE_NAME = "delta-exchange-mcp"
-METADATA_VERSION = 2
-SECRET_VERSION = 1
-DEFAULT_DIR = Path.home() / ".delta-exchange-mcp"
-DEFAULT_METADATA_NAME = "credentials.json"
-SUPPORTED_ENVIRONMENTS = frozenset({"india_prod", "india_testnet"})
-
-logger = logging.getLogger(__name__)
-
-_CREDENTIAL_NAMES = frozenset({"DELTA_API_KEY", "DELTA_API_SECRET"})
-_LOCK_TIMEOUT_SECONDS = 10.0
-_LOCK_POLL_SECONDS = 0.05
-_SECURE_KEYRING_TYPES = frozenset(
-    {
-        "keyring.backends.macOS.Keyring",
-        "keyring.backends.SecretService.Keyring",
-        "keyring.backends.libsecret.Keyring",
-        "keyring.backends.Windows.WinVaultKeyring",
-    }
+from delta_exchange_mcp.auth.backend import (
+    DEFAULT_DIR as DEFAULT_DIR,
+    DEFAULT_METADATA_NAME as DEFAULT_METADATA_NAME,
+    METADATA_VERSION as METADATA_VERSION,
+    SERVICE_NAME as SERVICE_NAME,
+    SECRET_VERSION as SECRET_VERSION,
+    SUPPORTED_ENVIRONMENTS as SUPPORTED_ENVIRONMENTS,
+    BackendOperationError as BackendOperationError,
+    BackendUnavailableError as BackendUnavailableError,
+    CredentialCorruptError as CredentialCorruptError,
+    CredentialMetadata as CredentialMetadata,
+    CredentialState as CredentialState,
+    CredentialStoreError as CredentialStoreError,
+    EnvironmentState,
+    FileMetadata as FileMetadata,
+    MemoryMetadata as MemoryMetadata,
+    MemorySecretBackend as MemorySecretBackend,
+    MetadataBackend,
+    MetadataError as MetadataError,
+    SecretBackend,
+    SystemKeyringBackend as SystemKeyringBackend,
+    _decode_secret,
+    _encode_secret,
+    _record_name,
+    default_metadata_path as default_metadata_path,
+    normalize_environment,
+    restore_record,
+)
+from delta_exchange_mcp.auth.migration import (
+    MigrationError as MigrationError,
+    MigrationResult as MigrationResult,
+    MigrationStatus as MigrationStatus,
+    migrate,
 )
 
-if os.name == "nt":
-    import msvcrt
-else:
-    import fcntl
-
-
-class CredentialState(StrEnum):
-    VERIFIED = "verified"
-    UNVERIFIED = "unverified"
+logger = logging.getLogger(__name__)
 
 
 class CredentialSource(StrEnum):
@@ -63,31 +53,7 @@ class CredentialSource(StrEnum):
     PROCESS = "process"
 
 
-class MigrationStatus(StrEnum):
-    ABSENT = "absent"
-    INCOMPLETE = "incomplete"
-    UNAVAILABLE = "unavailable"
-    MIGRATED = "migrated"
-    CONFLICT = "conflict"
-
-
-class CredentialStoreError(Exception):
-    pass
-
-
-class BackendUnavailableError(CredentialStoreError):
-    pass
-
-
-class BackendOperationError(CredentialStoreError):
-    pass
-
-
 class CredentialConflictError(CredentialStoreError):
-    pass
-
-
-class CredentialCorruptError(CredentialStoreError):
     pass
 
 
@@ -97,28 +63,6 @@ class CredentialActivationError(CredentialStoreError):
 
 class IncompleteCredentialError(CredentialStoreError):
     pass
-
-
-class MetadataError(CredentialStoreError):
-    pass
-
-
-class MigrationError(CredentialStoreError):
-    pass
-
-
-@dataclass(frozen=True)
-class CredentialMetadata:
-    environment: str
-    revision: int | None
-    generation: int
-    state: CredentialState | None
-    account_id: str
-    created_at: str
-    updated_at: str
-    validated_at: str
-    pending_revisions: tuple[int, ...]
-    reconnect_required: bool = False
 
 
 @dataclass(frozen=True)
@@ -147,309 +91,135 @@ class Credential:
 
 
 @dataclass(frozen=True)
-class MigrationResult:
-    status: MigrationStatus
+class _PreparedReplacement:
+    backend: SecretBackend
+    metadata: MetadataBackend
+    values: dict[str, EnvironmentState]
     environment: str
-    credential: Credential | None = None
+    had_previous: bool
+    previous: EnvironmentState
+    previous_credential: Credential | None
+    previous_payload: str | None
+    reserved: EnvironmentState
+    current: EnvironmentState
+    new_revision: int
+    new_name: str
+    new_payload: str
+    new_credential: Credential
+    activate: Callable[[Credential | None], None] | None
 
-
-@dataclass(frozen=True)
-class _EnvironmentState:
-    active_revision: int | None = None
-    next_revision: int = 1
-    generation: int = 0
-    state: CredentialState | None = None
-    account_id: str = ""
-    created_at: str = ""
-    updated_at: str = ""
-    validated_at: str = ""
-    pending_revisions: tuple[int, ...] = ()
-    preserved_records: tuple[str, ...] = ()
-    reconnect_required: bool = False
-
-    def metadata(self, environment: str) -> CredentialMetadata:
-        return CredentialMetadata(
-            environment=environment,
-            revision=self.active_revision,
-            generation=self.generation,
-            state=self.state,
-            account_id=self.account_id,
-            created_at=self.created_at,
-            updated_at=self.updated_at,
-            validated_at=self.validated_at,
-            pending_revisions=self.pending_revisions,
-            reconnect_required=self.reconnect_required,
-        )
-
-
-class SecretBackend(Protocol):
-    def get(self, name: str) -> str | None: ...
-
-    def set(self, name: str, value: str) -> None: ...
-
-    def delete(self, name: str) -> None: ...
-
-
-class MetadataBackend(Protocol):
-    @property
-    def namespace(self) -> str: ...
-
-    def lock(self) -> AbstractContextManager[None]: ...
-
-    def read(self) -> dict[str, _EnvironmentState]: ...
-
-    def write(self, values: dict[str, _EnvironmentState]) -> None: ...
-
-
-def default_metadata_path() -> Path:
-    """Return the path for non-secret credential metadata."""
-    return DEFAULT_DIR / DEFAULT_METADATA_NAME
-
-
-def _backend_name(backend: KeyringBackend) -> str:
-    kind = type(backend)
-    return f"{kind.__module__}.{kind.__qualname__}"
-
-
-class SystemKeyringBackend:
-    """Store secrets only in an approved operating-system keyring backend."""
-
-    def __init__(
-        self,
-        backend: KeyringBackend | None = None,
-        service_name: str = SERVICE_NAME,
-    ):
+    def write_new(self) -> None:
+        """Reserve, write, and read back the new secret record."""
+        self.values[self.environment] = self.reserved
+        self.metadata.write(self.values)
         try:
-            selected = backend if backend is not None else keyring.get_keyring()
-        except Exception as exc:
-            raise BackendUnavailableError(
-                f"could not discover a system keyring: {exc}"
-            ) from exc
-        name = _backend_name(selected)
-        try:
-            priority = selected.priority
-        except Exception as exc:
-            raise BackendUnavailableError(
-                f"keyring backend {name} is unavailable: {exc}"
-            ) from exc
-        if name not in _SECURE_KEYRING_TYPES or priority <= 0:
-            raise BackendUnavailableError(
-                f"keyring backend {name} is not an approved system credential store"
-            )
-        self._backend = selected
-        self._service_name = service_name
-
-    def get(self, name: str) -> str | None:
-        try:
-            return self._backend.get_password(self._service_name, name)
-        except Exception as exc:
-            raise BackendOperationError(
-                f"could not read {name} from the system store"
-            ) from exc
-
-    def set(self, name: str, value: str) -> None:
-        try:
-            self._backend.set_password(self._service_name, name, value)
-        except Exception as exc:
-            raise BackendOperationError(
-                f"could not write {name} to the system store"
-            ) from exc
-
-    def delete(self, name: str) -> None:
-        try:
-            self._backend.delete_password(self._service_name, name)
-        except PasswordDeleteError as exc:
-            if self.get(name) is None:
-                return
-            raise BackendOperationError(
-                f"could not delete {name} from the system store"
-            ) from exc
-        except Exception as exc:
-            raise BackendOperationError(
-                f"could not delete {name} from the system store"
-            ) from exc
-
-
-class MemorySecretBackend:
-    """Process-local fallback used when no approved system keyring is available."""
-
-    def __init__(self):
-        self._values: dict[str, str] = {}
-        self._lock = threading.RLock()
-
-    def get(self, name: str) -> str | None:
-        with self._lock:
-            return self._values.get(name)
-
-    def set(self, name: str, value: str) -> None:
-        with self._lock:
-            self._values[name] = value
-
-    def delete(self, name: str) -> None:
-        with self._lock:
-            self._values.pop(name, None)
-
-
-@contextmanager
-def _file_lock(target: Path) -> Iterator[tuple[int, Path]]:
-    lock_path = target.with_name(f".{target.name}.lock")
-    try:
-        target.parent.mkdir(mode=0o700, parents=True, exist_ok=True)
-        fd = os.open(lock_path, os.O_RDWR | os.O_CREAT, 0o600)
-    except OSError as exc:
-        raise MetadataError(f"could not open the lock for {target}: {exc}") from exc
-
-    deadline = time.monotonic() + _LOCK_TIMEOUT_SECONDS
-    locked = False
-    try:
-        if os.name == "nt":
-            if os.fstat(fd).st_size == 0:
-                os.write(fd, b"\0")
-            while not locked:
-                os.lseek(fd, 0, os.SEEK_SET)
-                try:
-                    msvcrt.locking(fd, msvcrt.LK_NBLCK, 1)
-                    locked = True
-                except OSError as exc:
-                    if time.monotonic() >= deadline:
-                        raise MetadataError(
-                            f"timed out waiting for {lock_path}"
-                        ) from exc
-                    time.sleep(_LOCK_POLL_SECONDS)
-        else:
-            while not locked:
-                try:
-                    fcntl.flock(fd, fcntl.LOCK_EX | fcntl.LOCK_NB)
-                    locked = True
-                except BlockingIOError as exc:
-                    if time.monotonic() >= deadline:
-                        raise MetadataError(
-                            f"timed out waiting for {lock_path}"
-                        ) from exc
-                    time.sleep(_LOCK_POLL_SECONDS)
-        yield fd, lock_path
-    finally:
-        if locked:
-            if os.name == "nt":
-                os.lseek(fd, 0, os.SEEK_SET)
-                msvcrt.locking(fd, msvcrt.LK_UNLCK, 1)
-            else:
-                fcntl.flock(fd, fcntl.LOCK_UN)
-        os.close(fd)
-
-
-class FileMetadata:
-    """Read and atomically replace the non-secret credential metadata file."""
-
-    def __init__(self, path: Path):
-        # The persistent lock exists before the first metadata write. Resolve its
-        # filesystem spelling so case aliases use the same namespace from startup.
-        with _file_lock(path.resolve()) as (fd, lock_path):
-            if os.name != "nt" and hasattr(fcntl, "F_GETPATH"):
-                raw = fcntl.fcntl(fd, fcntl.F_GETPATH, b"\0" * 1024)
-                lock_path = Path(os.fsdecode(raw.split(b"\0", 1)[0]))
-            else:
-                lock_path = lock_path.resolve()
-            self.path = lock_path.with_name(
-                lock_path.name.removeprefix(".").removesuffix(".lock")
-            )
-        self.namespace = hashlib.sha256(os.fsencode(self.path)).hexdigest()
-        self._thread_lock = threading.RLock()
-
-    @contextmanager
-    def lock(self) -> Iterator[None]:
-        with self._thread_lock:
-            with _file_lock(self.path):
-                yield
-
-    def read(self) -> dict[str, _EnvironmentState]:
-        try:
-            raw = self.path.read_text()
-        except FileNotFoundError:
-            return {}
-        except OSError as exc:
-            raise MetadataError(f"could not read {self.path}: {exc}") from exc
-        try:
-            document = json.loads(raw)
-        except json.JSONDecodeError as exc:
-            raise MetadataError(
-                f"credential metadata in {self.path} is not valid JSON"
-            ) from exc
-        if (
-            not isinstance(document, dict)
-            or type(document.get("version")) is not int
-            or document["version"] not in (1, METADATA_VERSION)
-        ):
-            raise MetadataError(
-                f"credential metadata in {self.path} has an unsupported version"
-            )
-        environments = document.get("environments")
-        if not isinstance(environments, dict):
-            raise MetadataError(
-                f"credential metadata in {self.path} has no environment map"
-            )
-        namespace = document.get("namespace")
-        if document["version"] == METADATA_VERSION and (
-            not isinstance(namespace, str)
-            or re.fullmatch(r"[0-9a-f]{64}", namespace) is None
-        ):
-            raise MetadataError(
-                f"credential metadata in {self.path} has an invalid namespace"
-            )
-        values = {
-            _normalize_environment(name): _state_from_json(value, self.path)
-            for name, value in environments.items()
-        }
-        if document["version"] == 1 or namespace != self.namespace:
-            # Old global names and copied metadata cannot establish ownership.
-            # Retain their record names for recovery, never read or delete them.
-            return {
-                environment: _detach_records(
-                    environment,
-                    state,
-                    namespace if document["version"] == METADATA_VERSION else None,
+            self.backend.set(self.new_name, self.new_payload)
+            if self.backend.get(self.new_name) != self.new_payload:
+                raise BackendOperationError(
+                    "the system store did not return credential revision "
+                    f"{self.new_revision} after writing it"
                 )
-                for environment, state in values.items()
-            }
-        return values
+        except Exception:
+            try:
+                self._discard_new()
+            except Exception as cleanup_exc:
+                raise CredentialStoreError(
+                    "credential write verification failed and cleanup remains pending"
+                ) from cleanup_exc
+            raise
 
-    def write(self, values: dict[str, _EnvironmentState]) -> None:
-        document = {
-            "version": METADATA_VERSION,
-            "namespace": self.namespace,
-            "environments": {
-                environment: asdict(state)
-                for environment, state in sorted(values.items())
-            },
-        }
-        body = f"{json.dumps(document, indent=2, sort_keys=True)}\n"
-        _atomic_replace(self.path, body, 0o600)
+    def publish(self) -> None:
+        """Publish the verified record as the active revision."""
+        self.values[self.environment] = self.current
+        try:
+            self.metadata.write(self.values)
+        except Exception:
+            self.values[self.environment] = self.reserved
+            try:
+                self._discard_new()
+            except Exception as cleanup_exc:
+                raise CredentialStoreError(
+                    "credential publication failed and cleanup remains pending"
+                ) from cleanup_exc
+            raise
 
+    def activate_new(self) -> None:
+        if self.activate is not None:
+            self.activate(self.new_credential)
 
-class MemoryMetadata:
-    """Keep metadata in this process when no secure keyring is available."""
+    def retire_previous(self) -> None:
+        revision = self.previous.active_revision
+        if revision is None:
+            return
+        self.backend.delete(
+            _record_name(self.metadata.namespace, self.environment, revision)
+        )
+        cleaned = replace_fields(
+            self.current,
+            pending_revisions=self.previous.pending_revisions,
+        )
+        self.values[self.environment] = cleaned
+        try:
+            self.metadata.write(self.values)
+        except Exception as exc:
+            self.values[self.environment] = self.current
+            logger.warning(
+                "activated credential revision %s for %s, but pending cleanup "
+                "metadata could not be cleared: %s",
+                self.current.active_revision,
+                self.environment,
+                exc,
+            )
 
-    def __init__(self):
-        self.namespace = secrets.token_hex(32)
-        self._values: dict[str, _EnvironmentState] = {}
-        self._lock = threading.RLock()
+    def restore_previous_record(self) -> None:
+        revision = self.previous.active_revision
+        if revision is None or self.previous_payload is None:
+            raise CredentialStoreError(
+                "credential retirement failed after removing the old record"
+            )
+        name = _record_name(self.metadata.namespace, self.environment, revision)
+        restore_record(self.backend, name, self.previous_payload)
 
-    @contextmanager
-    def lock(self) -> Iterator[None]:
-        with self._lock:
-            yield
+    def rollback(self) -> None:
+        """Restore the previous active record and remove the new record."""
+        rollback = replace_fields(
+            self.previous,
+            next_revision=self.current.next_revision,
+            pending_revisions=(
+                *self.previous.pending_revisions,
+                self.new_revision,
+            ),
+        )
+        self.values[self.environment] = rollback
+        self.metadata.write(self.values)
 
-    def read(self) -> dict[str, _EnvironmentState]:
-        with self._lock:
-            return dict(self._values)
+        activation_error: Exception | None = None
+        try:
+            if self.activate is not None:
+                self.activate(self.previous_credential)
+        except Exception as exc:
+            activation_error = exc
 
-    def write(self, values: dict[str, _EnvironmentState]) -> None:
-        with self._lock:
-            self._values = dict(values)
+        try:
+            self._discard_new()
+        except Exception as cleanup_exc:
+            raise CredentialStoreError(
+                "credential rollback failed and cleanup remains pending"
+            ) from cleanup_exc
+        if activation_error is not None:
+            raise activation_error
+
+    def _discard_new(self) -> None:
+        self.backend.delete(self.new_name)
+        if self.had_previous:
+            self.values[self.environment] = self.previous
+        else:
+            self.values.pop(self.environment, None)
+        self.metadata.write(self.values)
 
 
 class CredentialStore:
-    """Own versioned Delta credentials and their non-secret active pointers."""
+    """Coordinate versioned Delta credentials and their active pointers."""
 
     def __init__(
         self,
@@ -493,8 +263,8 @@ class CredentialStore:
 
     def metadata(self, environment: str) -> CredentialMetadata:
         """Read active revision and revocation state without reading a secret."""
-        env = _normalize_environment(environment)
-        state = self._metadata.read().get(env, _EnvironmentState())
+        env = normalize_environment(environment)
+        state = self._metadata.read().get(env, EnvironmentState())
         return state.metadata(env)
 
     def generation(self, environment: str) -> int:
@@ -503,7 +273,7 @@ class CredentialStore:
 
     def get(self, environment: str) -> Credential | None:
         """Read the active credential for an environment."""
-        env = _normalize_environment(environment)
+        env = normalize_environment(environment)
         with self._metadata.lock():
             values = self._metadata.read()
             self._cleanup_pending_locked(values, env)
@@ -515,7 +285,7 @@ class CredentialStore:
         environ: Mapping[str, str] | None = None,
     ) -> Credential | None:
         """Prefer a complete external process credential over the stored record."""
-        env = _normalize_environment(environment)
+        env = normalize_environment(environment)
         supplied = os.environ if environ is None else environ
         key = (supplied.get("DELTA_API_KEY") or "").strip()
         secret = (supplied.get("DELTA_API_SECRET") or "").strip()
@@ -556,7 +326,7 @@ class CredentialStore:
         activate: Callable[[Credential | None], None] | None = None,
     ) -> Credential:
         """Publish, activate, and retire one credential as a transaction."""
-        env = _normalize_environment(environment)
+        env = normalize_environment(environment)
         key = api_key.strip()
         secret = api_secret.strip()
         if not key or not secret:
@@ -565,7 +335,7 @@ class CredentialStore:
         with self._metadata.lock():
             values = self._metadata.read()
             self._cleanup_pending_locked(values, env)
-            return self._replace_locked(
+            transaction = self._prepare_replace_locked(
                 values,
                 env,
                 key,
@@ -576,6 +346,43 @@ class CredentialStore:
                 expected_generation=expected_generation,
                 activate=activate,
             )
+            transaction.write_new()
+            transaction.publish()
+            try:
+                transaction.activate_new()
+            except Exception as exc:
+                try:
+                    transaction.rollback()
+                except Exception as rollback_exc:
+                    raise CredentialStoreError(
+                        "credential activation failed and its transaction rollback also failed"
+                    ) from rollback_exc
+                if isinstance(exc, CredentialStoreError):
+                    raise
+                raise CredentialActivationError(
+                    f"could not activate credential revision {transaction.new_revision}"
+                ) from exc
+
+            try:
+                transaction.retire_previous()
+            except Exception as exc:
+                try:
+                    transaction.restore_previous_record()
+                except Exception as restore_exc:
+                    raise CredentialStoreError(
+                        "credential retirement failed and the old record could not be restored"
+                    ) from restore_exc
+                try:
+                    transaction.rollback()
+                except Exception as rollback_exc:
+                    raise CredentialStoreError(
+                        "credential retirement failed and its transaction rollback also failed"
+                    ) from rollback_exc
+                raise BackendOperationError(
+                    "could not retire credential revision "
+                    f"{transaction.previous.active_revision}"
+                ) from exc
+            return transaction.new_credential
 
     def delete(
         self,
@@ -585,7 +392,7 @@ class CredentialStore:
         expected_generation: int | None = None,
     ) -> bool:
         """Delete an active credential and advance its revocation generation."""
-        env = _normalize_environment(environment)
+        env = normalize_environment(environment)
         with self._metadata.lock():
             values = self._metadata.read()
             self._cleanup_pending_locked(values, env)
@@ -598,96 +405,11 @@ class CredentialStore:
 
     def migrate(self, config_path: Path) -> MigrationResult:
         """Move one complete legacy file credential into this store."""
-        with _file_lock(config_path):
-            if config_path.is_symlink():
-                return MigrationResult(MigrationStatus.UNAVAILABLE, "")
-            try:
-                original = config_path.read_text()
-            except FileNotFoundError:
-                return MigrationResult(MigrationStatus.ABSENT, "")
-            except OSError as exc:
-                raise MigrationError(f"could not read {config_path}: {exc}") from exc
+        return migrate(self, config_path)
 
-            parsed = dotenv_values(stream=io.StringIO(original))
-            key = (parsed.get("DELTA_API_KEY") or "").strip()
-            secret = (parsed.get("DELTA_API_SECRET") or "").strip()
-            environment = (parsed.get("DELTA_MCP_ENV") or "india_prod").strip().lower()
-            env = _normalize_environment(environment)
-            if not key and not secret:
-                return MigrationResult(MigrationStatus.ABSENT, env)
-            if not key or not secret:
-                return MigrationResult(MigrationStatus.INCOMPLETE, env)
-            if not self.persistent:
-                return MigrationResult(MigrationStatus.UNAVAILABLE, env)
-
-            rewritten = _without_credentials(original)
-            staged = _stage_replacement(config_path, rewritten)
-            try:
-                with self._metadata.lock():
-                    values = self._metadata.read()
-                    self._cleanup_pending_locked(values, env)
-                    current = self._get_locked(env, values)
-                    if current is not None and (
-                        current.api_key != key or current.api_secret != secret
-                    ):
-                        return MigrationResult(
-                            MigrationStatus.CONFLICT,
-                            env,
-                            current,
-                        )
-
-                    expected_revision = current.revision if current is not None else 0
-                    state = (
-                        current.state
-                        if current is not None
-                        else CredentialState.UNVERIFIED
-                    )
-                    account_id = current.account_id if current is not None else ""
-
-                    def publish_config(_credential: Credential) -> None:
-                        nonlocal staged
-                        try:
-                            os.replace(staged, config_path)
-                        except OSError as exc:
-                            raise MigrationError(
-                                f"could not publish the migrated {config_path}: {exc}"
-                            ) from exc
-                        staged = None
-                        _sync_after_publication(config_path.parent)
-
-                    migrated = self._replace_locked(
-                        values,
-                        env,
-                        key,
-                        secret,
-                        state=state,
-                        account_id=account_id,
-                        expected_revision=expected_revision,
-                        expected_generation=values.get(
-                            env,
-                            _EnvironmentState(),
-                        ).generation,
-                        publish=publish_config,
-                        keep_active_on_retirement_failure=True,
-                    )
-                    return MigrationResult(MigrationStatus.MIGRATED, env, migrated)
-            finally:
-                if staged is not None:
-                    staged.unlink(missing_ok=True)
-
-    def _get_locked(
+    def _prepare_replace_locked(
         self,
-        environment: str,
-        values: dict[str, _EnvironmentState],
-    ) -> Credential | None:
-        return self._credential(
-            environment,
-            values.get(environment, _EnvironmentState()),
-        )
-
-    def _replace_locked(
-        self,
-        values: dict[str, _EnvironmentState],
+        values: dict[str, EnvironmentState],
         environment: str,
         api_key: str,
         api_secret: str,
@@ -697,28 +419,41 @@ class CredentialStore:
         expected_revision: int | None,
         expected_generation: int | None,
         activate: Callable[[Credential | None], None] | None = None,
-        publish: Callable[[Credential], None] | None = None,
-        keep_active_on_retirement_failure: bool = False,
-    ) -> Credential:
+    ) -> _PreparedReplacement:
         had_previous = environment in values
-        previous = values.get(environment, _EnvironmentState())
+        previous = values.get(environment, EnvironmentState())
         _check_revision(previous, expected_revision)
         _check_generation(previous, expected_generation)
-        previous_credential = self._credential(environment, previous)
+
         previous_payload = None
+        previous_credential = None
         if previous.active_revision is not None:
-            previous_payload = self._backend.get(
-                self._record_name(environment, previous.active_revision)
+            previous_name = _record_name(
+                self._metadata.namespace,
+                environment,
+                previous.active_revision,
             )
+            previous_payload = self._backend.get(previous_name)
             if previous_payload is None:
                 raise CredentialCorruptError(
-                    f"credential metadata points to missing revision "
+                    "credential metadata points to missing revision "
                     f"{previous.active_revision} for {environment}"
                 )
+            previous_credential = _credential_from_payload(
+                environment,
+                previous,
+                previous_payload,
+                self.source,
+            )
 
         revision = previous.next_revision
         now = _now()
-        current = _EnvironmentState(
+        reserved = replace_fields(
+            previous,
+            next_revision=revision + 1,
+            pending_revisions=(*previous.pending_revisions, revision),
+        )
+        current = EnvironmentState(
             active_revision=revision,
             next_revision=revision + 1,
             generation=previous.generation + 1,
@@ -737,137 +472,65 @@ class CredentialStore:
             ),
             preserved_records=previous.preserved_records,
         )
-        name = self._record_name(environment, revision)
-        payload = _encode_secret(api_key, api_secret)
-        self._backend.set(name, payload)
-        if self._backend.get(name) != payload:
-            self._delete_new(name)
-            raise BackendOperationError(
-                f"the system store did not return credential revision {revision} after writing it"
-            )
-
-        values[environment] = current
-        try:
-            self._metadata.write(values)
-        except Exception:
-            self._delete_new(name)
-            raise
-
-        try:
-            credential = self._credential(environment, current)
-            if credential is None:
-                raise CredentialCorruptError(
-                    "the new credential has no active revision"
-                )
-            if activate is not None:
-                activate(credential)
-        except Exception as exc:
-            self._rollback_replace(
-                values,
+        new_payload = _encode_secret(api_key, api_secret)
+        return _PreparedReplacement(
+            backend=self._backend,
+            metadata=self._metadata,
+            values=values,
+            environment=environment,
+            had_previous=had_previous,
+            previous=previous,
+            previous_credential=previous_credential,
+            previous_payload=previous_payload,
+            reserved=reserved,
+            current=current,
+            new_revision=revision,
+            new_name=_record_name(self._metadata.namespace, environment, revision),
+            new_payload=new_payload,
+            new_credential=_credential_from_payload(
                 environment,
-                previous,
-                had_previous,
-                name,
-                activate,
-                previous_credential,
-            )
-            if isinstance(exc, CredentialStoreError):
-                raise
-            raise CredentialActivationError(
-                f"could not activate credential revision {revision}"
-            ) from exc
-
-        try:
-            if publish is not None:
-                publish(credential)
-        except Exception:
-            self._rollback_replace(
-                values,
-                environment,
-                previous,
-                had_previous,
-                name,
-                activate,
-                previous_credential,
-            )
-            raise
-
-        if previous.active_revision is not None:
-            old_name = self._record_name(environment, previous.active_revision)
-            try:
-                self._backend.delete(old_name)
-            except Exception as exc:
-                if keep_active_on_retirement_failure:
-                    logger.warning(
-                        "could not retire inactive credential revision %s for %s "
-                        "after migration publication; cleanup remains pending: %s",
-                        previous.active_revision,
-                        environment,
-                        exc,
-                    )
-                    return credential
-                if previous_payload is None:
-                    raise CredentialStoreError(
-                        "credential retirement failed after removing the old record"
-                    ) from exc
-                try:
-                    self._restore_record(old_name, previous_payload)
-                except Exception as restore_exc:
-                    raise CredentialStoreError(
-                        "credential retirement failed and the old record could not be restored"
-                    ) from restore_exc
-                try:
-                    self._rollback_replace(
-                        values,
-                        environment,
-                        previous,
-                        had_previous,
-                        name,
-                        activate,
-                        previous_credential,
-                    )
-                except Exception as rollback_exc:
-                    raise CredentialStoreError(
-                        "credential retirement failed and its transaction rollback also failed"
-                    ) from rollback_exc
-                raise BackendOperationError(
-                    f"could not retire credential revision {previous.active_revision}"
-                ) from exc
-            cleaned = replace_fields(
                 current,
-                pending_revisions=previous.pending_revisions,
+                new_payload,
+                self.source,
+            ),
+            activate=activate,
+        )
+
+    def _get_locked(
+        self,
+        environment: str,
+        values: dict[str, EnvironmentState],
+    ) -> Credential | None:
+        state = values.get(environment, EnvironmentState())
+        revision = state.active_revision
+        if revision is None:
+            return None
+        payload = self._backend.get(
+            _record_name(self._metadata.namespace, environment, revision)
+        )
+        if payload is None:
+            raise CredentialCorruptError(
+                f"credential metadata points to missing revision {revision} for {environment}"
             )
-            values[environment] = cleaned
-            try:
-                self._metadata.write(values)
-            except Exception as exc:
-                values[environment] = current
-                logger.warning(
-                    "activated credential revision %s for %s, but pending cleanup "
-                    "metadata could not be cleared: %s",
-                    revision,
-                    environment,
-                    exc,
-                )
-        return credential
+        return _credential_from_payload(environment, state, payload, self.source)
 
     def _delete_locked(
         self,
-        values: dict[str, _EnvironmentState],
+        values: dict[str, EnvironmentState],
         environment: str,
         expected_revision: int | None,
         expected_generation: int | None,
     ) -> bool:
-        previous = values.get(environment, _EnvironmentState())
+        previous = values.get(environment, EnvironmentState())
         _check_revision(previous, expected_revision)
         _check_generation(previous, expected_generation)
         if previous.active_revision is None:
             return False
         revision = previous.active_revision
-        name = self._record_name(environment, revision)
+        name = _record_name(self._metadata.namespace, environment, revision)
         payload = self._backend.get(name)
 
-        tombstone = _EnvironmentState(
+        tombstone = EnvironmentState(
             active_revision=None,
             next_revision=previous.next_revision,
             generation=previous.generation + 1,
@@ -893,7 +556,7 @@ class CredentialStore:
             self._backend.delete(name)
         except Exception as exc:
             try:
-                self._restore_record(name, payload)
+                restore_record(self._backend, name, payload)
             except Exception as restore_exc:
                 raise CredentialStoreError(
                     "credential deletion failed after the disconnect was published; "
@@ -941,16 +604,16 @@ class CredentialStore:
 
     def _cleanup_pending_locked(
         self,
-        values: dict[str, _EnvironmentState],
+        values: dict[str, EnvironmentState],
         environment: str,
     ) -> None:
-        previous = values.get(environment, _EnvironmentState())
+        previous = values.get(environment, EnvironmentState())
         if not previous.pending_revisions:
             return
 
         remaining: list[int] = []
         for revision in previous.pending_revisions:
-            name = self._record_name(environment, revision)
+            name = _record_name(self._metadata.namespace, environment, revision)
             try:
                 self._backend.delete(name)
             except Exception as exc:
@@ -977,10 +640,7 @@ class CredentialStore:
         pending_revisions = tuple(remaining)
         if pending_revisions == previous.pending_revisions:
             return
-        current = replace_fields(
-            previous,
-            pending_revisions=pending_revisions,
-        )
+        current = replace_fields(previous, pending_revisions=pending_revisions)
         values[environment] = current
         try:
             self._metadata.write(values)
@@ -1015,140 +675,35 @@ class CredentialStore:
                     self._process_generations.get(environment, 0) + 1
                 )
 
-    def _credential(
-        self, environment: str, state: _EnvironmentState
-    ) -> Credential | None:
-        revision = state.active_revision
-        if revision is None:
-            return None
-        payload = self._backend.get(self._record_name(environment, revision))
-        if payload is None:
-            raise CredentialCorruptError(
-                f"credential metadata points to missing revision {revision} for {environment}"
-            )
-        key, secret = _decode_secret(payload)
-        if state.state is None:
-            raise CredentialCorruptError(
-                f"credential revision {revision} for {environment} has no validation state"
-            )
-        return Credential(
-            environment=environment,
-            revision=revision,
-            generation=state.generation,
-            state=state.state,
-            source=self.source,
-            account_id=state.account_id,
-            created_at=state.created_at,
-            updated_at=state.updated_at,
-            validated_at=state.validated_at,
-            api_key=key,
-            api_secret=secret,
-        )
 
-    def _record_name(self, environment: str, revision: int) -> str:
-        return _record_name(self._metadata.namespace, environment, revision)
-
-    def _delete_new(self, name: str) -> None:
-        try:
-            self._backend.delete(name)
-        except Exception as exc:
-            raise CredentialStoreError(
-                f"credential write failed and temporary record {name} could not be removed"
-            ) from exc
-
-    def _restore_record(self, name: str, payload: str) -> None:
-        if self._backend.get(name) == payload:
-            return
-        self._backend.set(name, payload)
-        if self._backend.get(name) != payload:
-            raise CredentialStoreError(
-                f"credential record {name} could not be restored"
-            )
-
-    def _rollback_replace(
-        self,
-        values: dict[str, _EnvironmentState],
-        environment: str,
-        previous: _EnvironmentState,
-        had_previous: bool,
-        new_name: str,
-        activate: Callable[[Credential | None], None] | None,
-        previous_credential: Credential | None,
-    ) -> None:
-        if had_previous:
-            values[environment] = previous
-        else:
-            values.pop(environment, None)
-        self._metadata.write(values)
-        try:
-            if activate is not None:
-                activate(previous_credential)
-        finally:
-            self._delete_new(new_name)
-
-
-def _normalize_environment(environment: str) -> str:
-    value = environment.strip().lower()
-    if value not in SUPPORTED_ENVIRONMENTS:
-        raise ValueError(
-            f"Delta environment must be one of {sorted(SUPPORTED_ENVIRONMENTS)}, "
-            f"got {environment!r}"
-        )
-    return value
-
-
-def _record_name(namespace: str, environment: str, revision: int) -> str:
-    return f"credential:{namespace}:{environment}:{revision}"
-
-
-def _detach_records(
-    environment: str, state: _EnvironmentState, namespace: str | None
-) -> _EnvironmentState:
-    revisions = (
-        *state.pending_revisions,
-        *((state.active_revision,) if state.active_revision is not None else ()),
-    )
-    records = tuple(
-        _record_name(namespace, environment, revision)
-        if namespace is not None
-        else f"credential:{environment}:{revision}"
-        for revision in revisions
-    )
-    return _EnvironmentState(
-        next_revision=state.next_revision,
-        generation=state.generation + 1,
-        updated_at=state.updated_at,
-        preserved_records=tuple(dict.fromkeys((*state.preserved_records, *records))),
-        reconnect_required=state.active_revision is not None
-        or state.reconnect_required,
-    )
-
-
-def _encode_secret(api_key: str, api_secret: str) -> str:
-    return json.dumps(
-        {"version": SECRET_VERSION, "api_key": api_key, "api_secret": api_secret},
-        separators=(",", ":"),
-        sort_keys=True,
-    )
-
-
-def _decode_secret(payload: str) -> tuple[str, str]:
-    try:
-        value = json.loads(payload)
-    except json.JSONDecodeError as exc:
-        raise CredentialCorruptError("the credential record is not valid JSON") from exc
-    if not isinstance(value, dict) or value.get("version") != SECRET_VERSION:
-        raise CredentialCorruptError("the credential record has an unsupported version")
-    key = value.get("api_key")
-    secret = value.get("api_secret")
-    if not isinstance(key, str) or not key or not isinstance(secret, str) or not secret:
+def _credential_from_payload(
+    environment: str,
+    state: EnvironmentState,
+    payload: str,
+    source: CredentialSource,
+) -> Credential:
+    key, secret = _decode_secret(payload)
+    if state.state is None or state.active_revision is None:
         raise CredentialCorruptError(
-            "the credential record has no complete credential pair"
+            f"credential revision {state.active_revision} for {environment} "
+            "has no validation state"
         )
-    return key, secret
+    return Credential(
+        environment=environment,
+        revision=state.active_revision,
+        generation=state.generation,
+        state=state.state,
+        source=source,
+        account_id=state.account_id,
+        created_at=state.created_at,
+        updated_at=state.updated_at,
+        validated_at=state.validated_at,
+        api_key=key,
+        api_secret=secret,
+    )
 
 
-def _check_revision(state: _EnvironmentState, expected: int | None) -> None:
+def _check_revision(state: EnvironmentState, expected: int | None) -> None:
     if expected is None:
         return
     current = state.active_revision or 0
@@ -1158,7 +713,7 @@ def _check_revision(state: _EnvironmentState, expected: int | None) -> None:
         )
 
 
-def _check_generation(state: _EnvironmentState, expected: int | None) -> None:
+def _check_generation(state: EnvironmentState, expected: int | None) -> None:
     if expected is None:
         return
     if state.generation != expected:
@@ -1168,174 +723,5 @@ def _check_generation(state: _EnvironmentState, expected: int | None) -> None:
         )
 
 
-def _state_from_json(value: object, path: Path) -> _EnvironmentState:
-    if not isinstance(value, dict):
-        raise MetadataError(
-            f"credential metadata in {path} has an invalid environment entry"
-        )
-    try:
-        active = value["active_revision"]
-        next_revision = value["next_revision"]
-        generation = value["generation"]
-        raw_state = value["state"]
-        state = CredentialState(raw_state) if raw_state is not None else None
-        account_id = value["account_id"]
-        created_at = value["created_at"]
-        updated_at = value["updated_at"]
-        validated_at = value["validated_at"]
-        raw_pending = value.get("pending_revisions", [])
-        raw_preserved = value.get("preserved_records", [])
-        reconnect_required = value.get("reconnect_required", False)
-    except (KeyError, TypeError, ValueError) as exc:
-        raise MetadataError(
-            f"credential metadata in {path} has an invalid entry"
-        ) from exc
-    if active is not None and (type(active) is not int or active < 1):
-        raise MetadataError(
-            f"credential metadata in {path} has an invalid active revision"
-        )
-    if type(next_revision) is not int or next_revision < 1:
-        raise MetadataError(
-            f"credential metadata in {path} has an invalid next revision"
-        )
-    if type(generation) is not int or generation < 0:
-        raise MetadataError(f"credential metadata in {path} has an invalid generation")
-    if type(reconnect_required) is not bool or (
-        reconnect_required and active is not None
-    ):
-        raise MetadataError(
-            f"credential metadata in {path} has an invalid reconnect state"
-        )
-    if not isinstance(raw_preserved, list) or any(
-        not isinstance(name, str)
-        or re.fullmatch(
-            r"credential:(?:[0-9a-f]{64}:)?india_(?:prod|testnet):[1-9][0-9]*", name
-        )
-        is None
-        for name in raw_preserved
-    ):
-        raise MetadataError(
-            f"credential metadata in {path} has invalid preserved records"
-        )
-    if not isinstance(raw_pending, list) or any(
-        type(revision) is not int or revision < 1 for revision in raw_pending
-    ):
-        raise MetadataError(
-            f"credential metadata in {path} has invalid pending revisions"
-        )
-    pending_revisions = tuple(raw_pending)
-    if len(set(pending_revisions)) != len(pending_revisions):
-        raise MetadataError(
-            f"credential metadata in {path} has duplicate pending revisions"
-        )
-    if any(revision >= next_revision for revision in pending_revisions):
-        raise MetadataError(
-            f"credential metadata in {path} has a pending revision that was not issued"
-        )
-    if active in pending_revisions:
-        raise MetadataError(
-            f"credential metadata in {path} marks its active revision for cleanup"
-        )
-    if active is not None and next_revision <= active:
-        raise MetadataError(
-            f"credential metadata in {path} would reuse an active revision"
-        )
-    if active is not None and state is None:
-        raise MetadataError(f"credential metadata in {path} has no validation state")
-    if not all(
-        isinstance(item, str)
-        for item in (account_id, created_at, updated_at, validated_at)
-    ):
-        raise MetadataError(f"credential metadata in {path} has an invalid text field")
-    return _EnvironmentState(
-        active_revision=active,
-        next_revision=next_revision,
-        generation=generation,
-        state=state,
-        account_id=account_id,
-        created_at=created_at,
-        updated_at=updated_at,
-        validated_at=validated_at,
-        pending_revisions=pending_revisions,
-        preserved_records=tuple(raw_preserved),
-        reconnect_required=reconnect_required,
-    )
-
-
 def _now() -> str:
     return datetime.now(UTC).isoformat().replace("+00:00", "Z")
-
-
-def _atomic_replace(path: Path, body: str, mode: int) -> None:
-    staged: Path | None = None
-    try:
-        path.parent.mkdir(mode=0o700, parents=True, exist_ok=True)
-        fd, name = tempfile.mkstemp(
-            dir=path.parent, prefix=f".{path.name}-", suffix=".tmp"
-        )
-        staged = Path(name)
-        with os.fdopen(fd, "w") as handle:
-            handle.write(body)
-            handle.flush()
-            os.fsync(handle.fileno())
-        os.chmod(staged, mode)
-        os.replace(staged, path)
-        staged = None
-        _sync_after_publication(path.parent)
-    except OSError as exc:
-        raise MetadataError(f"could not write {path}: {exc}") from exc
-    finally:
-        if staged is not None:
-            staged.unlink(missing_ok=True)
-
-
-def _stage_replacement(path: Path, body: str) -> Path:
-    staged: Path | None = None
-    try:
-        mode = stat.S_IMODE(path.stat().st_mode) & 0o700
-        fd, name = tempfile.mkstemp(
-            dir=path.parent, prefix=f".{path.name}-", suffix=".tmp"
-        )
-        staged = Path(name)
-        with os.fdopen(fd, "w") as handle:
-            handle.write(body)
-            handle.flush()
-            os.fsync(handle.fileno())
-        os.chmod(staged, mode)
-        return staged
-    except OSError as exc:
-        if staged is not None:
-            staged.unlink(missing_ok=True)
-        raise MigrationError(
-            f"could not stage the migration for {path}: {exc}"
-        ) from exc
-
-
-def _without_credentials(body: str) -> str:
-    return "".join(
-        binding.original.string
-        for binding in parse_stream(io.StringIO(body))
-        if binding.key not in _CREDENTIAL_NAMES
-    )
-
-
-def _sync_directory(path: Path) -> None:
-    if os.name == "nt":
-        return
-    try:
-        fd = os.open(path, os.O_RDONLY)
-    except OSError:
-        return
-    try:
-        os.fsync(fd)
-    finally:
-        os.close(fd)
-
-
-def _sync_after_publication(path: Path) -> None:
-    try:
-        _sync_directory(path)
-    except OSError as exc:
-        # The replacement is already visible. A durability warning must not make
-        # callers delete the secret that the published pointer now names.
-        logger.warning("could not sync published directory %s: %s", path, exc)
