+++ tests/test_auth_store.py
@@ -1,137 +1,31 @@
 import json
-import os
 import threading
-from contextlib import contextmanager
 
 import pytest
-from keyring.backend import KeyringBackend
 
+from delta_exchange_mcp.auth import backend as auth_backend
 from delta_exchange_mcp.auth import store as auth_store
-from delta_exchange_mcp.auth.store import (
+from delta_exchange_mcp.auth.backend import (
     BackendOperationError,
-    BackendUnavailableError,
+    CredentialCorruptError,
+    CredentialState,
+    CredentialStoreError,
+    FileMetadata,
+)
+from delta_exchange_mcp.auth.store import (
     CredentialActivationError,
     CredentialConflictError,
-    CredentialCorruptError,
     CredentialSource,
-    CredentialState,
     CredentialStore,
-    FileMetadata,
     IncompleteCredentialError,
-    MemoryMetadata,
-    MigrationStatus,
-    SystemKeyringBackend,
 )
-
-
-class SimulatedProcessDeath(BaseException):
-    pass
-
-
-class FakeSecretBackend:
-    def __init__(self):
-        self.values: dict[str, str] = {}
-        self.reads: list[str] = []
-        self.deleted: list[str] = []
-        self.mismatch: set[str] = set()
-        self.fail_delete: set[str] = set()
-        self.fail_delete_after: set[str] = set()
-        self.crash_delete: set[str] = set()
-        self.crash_delete_after: set[str] = set()
-        self._lock = threading.RLock()
-
-    def get(self, name):
-        with self._lock:
-            self.reads.append(name)
-            if name in self.mismatch and name in self.values:
-                return "wrong readback"
-            return self.values.get(name)
-
-    def set(self, name, value):
-        with self._lock:
-            self.values[name] = value
-
-    def delete(self, name):
-        with self._lock:
-            if name in self.fail_delete:
-                self.fail_delete.remove(name)
-                raise BackendOperationError("delete failed")
-            if name in self.crash_delete:
-                self.crash_delete.remove(name)
-                raise SimulatedProcessDeath
-            self.values.pop(name, None)
-            self.deleted.append(name)
-            if name in self.fail_delete_after:
-                self.fail_delete_after.remove(name)
-                raise BackendOperationError("delete failed after removal")
-            if name in self.crash_delete_after:
-                self.crash_delete_after.remove(name)
-                raise SimulatedProcessDeath
-
-
-class FailingMetadata:
-    def __init__(self):
-        self.inner = MemoryMetadata()
-        self.fail_next_write = False
-
-    @property
-    def namespace(self):
-        return self.inner.namespace
-
-    @contextmanager
-    def lock(self):
-        with self.inner.lock():
-            yield
-
-    def read(self):
-        return self.inner.read()
-
-    def write(self, values):
-        if self.fail_next_write:
-            self.fail_next_write = False
-            raise auth_store.MetadataError("metadata write failed")
-        self.inner.write(values)
-
-
-def make_store(tmp_path, backend=None):
-    secret_backend = backend or FakeSecretBackend()
-    return (
-        CredentialStore(
-            secret_backend,
-            FileMetadata(tmp_path / "credentials.json"),
-            CredentialSource.OS_STORE,
-        ),
-        secret_backend,
-    )
-
-
-def record_name(credentials, revision=1, environment="india_prod"):
-    return f"credential:{credentials._metadata.namespace}:{environment}:{revision}"
-
-
-def legacy_metadata(tmp_path, backend):
-    path = tmp_path / "credentials.json"
-    document = {
-        "version": 1,
-        "environments": {
-            "india_prod": {
-                "active_revision": 2,
-                "next_revision": 3,
-                "generation": 2,
-                "state": "verified",
-                "account_id": "old-account",
-                "created_at": "2026-08-01T00:00:00Z",
-                "updated_at": "2026-08-01T00:00:00Z",
-                "validated_at": "2026-08-01T00:00:00Z",
-                "pending_revisions": [1],
-            }
-        },
-    }
-    path.write_text(json.dumps(document))
-    for revision in (1, 2):
-        # These global records could belong to a different metadata folder.
-        backend.values[f"credential:india_prod:{revision}"] = "unowned secret record"
-    return path
+from tests.credentials import (
+    FailingMetadata,
+    FakeSecretBackend,
+    SimulatedProcessDeath,
+    make_store,
+    record_name,
+)
 
 
 def test_replace_publishes_one_active_revision_without_secrets_in_metadata(tmp_path):
@@ -208,7 +102,7 @@ def test_directory_sync_failure_after_metadata_publication_keeps_new_record_acti
         directory_sync_failed = True
         raise OSError(f"could not sync {path}")
 
-    monkeypatch.setattr(auth_store, "_sync_directory", fail_directory_sync)
+    monkeypatch.setattr(auth_backend, "_sync_directory", fail_directory_sync)
 
     second = credentials.replace(
         "india_prod",
@@ -221,267 +115,6 @@ def test_directory_sync_failure_after_metadata_publication_keeps_new_record_acti
     assert directory_sync_failed is True
     assert credentials.get("india_prod") == second
     assert set(backend.values) == {record_name(credentials, 2)}
-
-
-def test_each_environment_has_its_own_active_record(tmp_path):
-    credentials, backend = make_store(tmp_path)
-
-    prod = credentials.replace("india_prod", "prod-key", "prod-secret")
-    testnet = credentials.replace("india_testnet", "test-key", "test-secret")
-
-    assert (prod.revision, testnet.revision) == (1, 1)
-    assert set(backend.values) == {
-        record_name(credentials, 1),
-        record_name(credentials, 1, "india_testnet"),
-    }
-
-
-@pytest.mark.parametrize("environment", ["india_prod", "india_testnet"])
-def test_separate_metadata_folders_keep_their_own_credentials(tmp_path, environment):
-    backend = FakeSecretBackend()
-    first, _ = make_store(tmp_path / "first", backend)
-    second, _ = make_store(tmp_path / "second", backend)
-
-    first_saved = first.replace(
-        environment, "first-key", "first-secret", account_id="first"
-    )
-    second_saved = second.replace(
-        environment, "second-key", "second-secret", account_id="second"
-    )
-
-    assert first.get(environment) == first_saved
-    assert second.get(environment) == second_saved
-    assert len(backend.values) == 2
-
-
-@pytest.mark.parametrize("operation", ["rotate", "disconnect"])
-def test_other_metadata_folders_cannot_replace_or_delete_a_credential(
-    tmp_path, operation
-):
-    backend = FakeSecretBackend()
-    first, _ = make_store(tmp_path / "first", backend)
-    second, _ = make_store(tmp_path / "second", backend)
-    first_saved = first.replace("india_prod", "first-key", "first-secret")
-    second.replace("india_prod", "second-key", "second-secret")
-
-    if operation == "rotate":
-        second.replace("india_prod", "rotated-key", "rotated-secret")
-    else:
-        second.delete("india_prod")
-
-    assert first.get("india_prod") == first_saved
-
-
-def test_metadata_files_in_the_same_folder_have_separate_records(tmp_path):
-    backend = FakeSecretBackend()
-    first, _ = make_store(tmp_path, backend)
-    second = CredentialStore(
-        backend, FileMetadata(tmp_path / "other.json"), CredentialSource.OS_STORE
-    )
-    saved = first.replace("india_prod", "first-key", "first-secret")
-    second.replace("india_prod", "second-key", "second-secret")
-
-    assert first.get("india_prod") == saved
-    assert len(backend.values) == 2
-
-
-@pytest.mark.parametrize("alias_kind", ["relative", "symlink", "case"])
-def test_metadata_path_aliases_share_one_record_and_revision_lock(tmp_path, alias_kind):
-    directory = tmp_path / "original"
-    directory.mkdir()
-    if alias_kind == "symlink":
-        alias = tmp_path / "alias"
-        try:
-            alias.symlink_to(directory, target_is_directory=True)
-        except OSError:
-            pytest.skip("directory symlinks are unavailable on this system")
-    elif alias_kind == "case":
-        alias = tmp_path / "ORIGINAL"
-        if not alias.exists():
-            pytest.skip("this filesystem has case-sensitive directory names")
-    else:
-        alias = directory / ".." / "original"
-    backend = FakeSecretBackend()
-    first, _ = make_store(directory, backend)
-    second, _ = make_store(alias, backend)
-    saved = first.replace("india_prod", "key", "secret")
-
-    assert second.get("india_prod") == saved
-    updated = second.replace("india_prod", "new-key", "secret", expected_revision=1)
-    with pytest.raises(CredentialConflictError):
-        first.replace("india_prod", "stale-key", "secret", expected_revision=1)
-    assert first.get("india_prod") == updated
-    assert len(backend.values) == 1
-
-
-def test_record_names_follow_filesystem_case_rules_before_the_first_save(tmp_path):
-    backend = FakeSecretBackend()
-    first_path = tmp_path / "credentials.json"
-    second_path = tmp_path / "CREDENTIALS.JSON"
-    first = CredentialStore(
-        backend, FileMetadata(first_path), CredentialSource.OS_STORE
-    )
-    second = CredentialStore(
-        backend, FileMetadata(second_path), CredentialSource.OS_STORE
-    )
-    saved = first.replace("india_prod", "first-key", "secret")
-
-    if second_path.exists() and first_path.samefile(second_path):
-        assert second.get("india_prod") == saved
-        updated = second.replace(
-            "india_prod", "second-key", "secret", expected_revision=1
-        )
-        assert first.get("india_prod") == updated
-        assert len(backend.values) == 1
-    else:
-        assert second.get("india_prod") is None
-        second.replace("india_prod", "second-key", "secret", expected_revision=0)
-        assert first.get("india_prod") == saved
-        assert len(backend.values) == 2
-
-
-def test_memory_metadata_instances_do_not_share_secret_names():
-    backend = FakeSecretBackend()
-    first = CredentialStore(backend, MemoryMetadata(), CredentialSource.MEMORY)
-    second = CredentialStore(backend, MemoryMetadata(), CredentialSource.MEMORY)
-    saved = first.replace("india_prod", "first-key", "secret")
-    second.replace("india_prod", "second-key", "secret")
-    second.delete("india_prod")
-
-    assert first.get("india_prod") == saved
-
-
-def test_legacy_metadata_requires_reconnect_without_accessing_old_os_records(
-    tmp_path, monkeypatch
-):
-    backend = FakeSecretBackend()
-    path = legacy_metadata(tmp_path, backend)
-    original = path.read_bytes()
-    original_records = dict(backend.values)
-    monkeypatch.setattr(auth_store, "SystemKeyringBackend", lambda: backend)
-
-    for _ in range(2):
-        credentials = CredentialStore.open(path)
-        assert credentials.get("india_prod") is None
-        metadata = credentials.metadata("india_prod")
-        assert metadata.revision is None
-        assert metadata.generation == 3
-        assert metadata.reconnect_required is True
-        assert metadata.pending_revisions == ()
-        assert metadata.account_id == ""
-        assert credentials.delete("india_prod") is False
-
-    assert path.read_bytes() == original
-    assert backend.values == original_records
-    assert backend.reads == []
-    assert backend.deleted == []
-
-
-@pytest.mark.parametrize("operation", ["replace", "migrate"])
-def test_reconnect_keeps_unowned_records_out_of_rotation_and_cleanup(
-    tmp_path, operation
-):
-    backend = FakeSecretBackend()
-    path = legacy_metadata(tmp_path, backend)
-    original_records = dict(backend.values)
-    credentials, _ = make_store(tmp_path, backend)
-
-    if operation == "migrate":
-        config_path = tmp_path / "config.env"
-        config_path.write_text(
-            "DELTA_API_KEY=new-key\nDELTA_API_SECRET=new-secret\nDELTA_MCP_MODE=trade\n"
-        )
-        migrated = credentials.migrate(config_path)
-        assert migrated.status is MigrationStatus.MIGRATED
-        assert config_path.read_text() == "DELTA_MCP_MODE=trade\n"
-    else:
-        credentials.replace(
-            "india_prod",
-            "new-key",
-            "new-secret",
-            expected_revision=0,
-            expected_generation=3,
-        )
-    saved = credentials.get("india_prod")
-    assert saved is not None
-    assert (saved.api_key, saved.api_secret) == ("new-key", "new-secret")
-    assert (saved.revision, saved.generation) == (3, 4)
-    assert credentials.metadata("india_prod").reconnect_required is False
-    credentials.replace("india_prod", "rotated-key", "secret")
-    assert credentials.delete("india_prod") is True
-
-    assert backend.values == original_records
-    assert set(original_records).isdisjoint(backend.reads)
-    assert set(original_records).isdisjoint(backend.deleted)
-    document = json.loads(path.read_text())
-    assert document["version"] == 2
-    metadata = document["environments"]["india_prod"]
-    assert set(metadata["preserved_records"]) == set(original_records)
-    assert metadata["pending_revisions"] == []
-    assert metadata["reconnect_required"] is False
-
-
-def test_reconnect_activation_failure_preserves_recovery_metadata(tmp_path):
-    backend = FakeSecretBackend()
-    path = legacy_metadata(tmp_path, backend)
-    original_records = dict(backend.values)
-    credentials, _ = make_store(tmp_path, backend)
-
-    def fail_activation(credential):
-        if credential is not None:
-            raise RuntimeError("cannot activate")
-
-    with pytest.raises(CredentialActivationError):
-        credentials.replace("india_prod", "new-key", "secret", activate=fail_activation)
-
-    assert credentials.get("india_prod") is None
-    assert credentials.metadata("india_prod").reconnect_required is True
-    assert backend.values == original_records
-    metadata = json.loads(path.read_text())["environments"]["india_prod"]
-    assert set(metadata["preserved_records"]) == set(original_records)
-
-
-def test_copied_metadata_cannot_read_or_clean_the_original_records(tmp_path):
-    backend = FakeSecretBackend()
-    original, _ = make_store(tmp_path / "original", backend)
-    original.replace("india_prod", "old-key", "secret")
-    backend.crash_delete.add(record_name(original, 1))
-    with pytest.raises(SimulatedProcessDeath):
-        original.replace("india_prod", "current-key", "secret")
-    source_records = dict(backend.values)
-    copied_path = tmp_path / "copy" / "credentials.json"
-    copied_path.parent.mkdir()
-    copied_path.write_bytes((tmp_path / "original" / "credentials.json").read_bytes())
-    copied, _ = make_store(copied_path.parent, backend)
-    backend.reads.clear()
-
-    assert copied.get("india_prod") is None
-    assert copied.metadata("india_prod").reconnect_required is True
-    copied.replace("india_prod", "copy-key", "secret")
-    copied.delete("india_prod")
-
-    assert backend.values == source_records
-    assert not set(source_records).intersection(backend.reads)
-    preserved = json.loads(copied_path.read_text())["environments"]["india_prod"]
-    assert set(preserved["preserved_records"]) == set(source_records)
-    assert preserved["pending_revisions"] == []
-
-
-@pytest.mark.parametrize("namespace", [None, 42, "", "a" * 63, "A" * 64])
-def test_invalid_namespace_never_reads_or_deletes_credentials(tmp_path, namespace):
-    credentials, backend = make_store(tmp_path)
-    credentials.replace("india_prod", "key", "secret")
-    path = tmp_path / "credentials.json"
-    document = json.loads(path.read_text())
-    document["namespace"] = namespace
-    path.write_text(json.dumps(document))
-    backend.reads.clear()
-    backend.deleted.clear()
-
-    with pytest.raises(auth_store.MetadataError, match="namespace"):
-        credentials.get("india_prod")
-    assert backend.reads == []
-    assert backend.deleted == []
 
 
 def test_a_stale_replace_never_writes_a_new_secret(tmp_path):
@@ -506,6 +139,79 @@ def test_a_failed_readback_removes_the_new_record_and_keeps_metadata_empty(tmp_p
 
     assert backend.values == {}
     assert credentials.metadata("india_prod").revision is None
+
+
+def test_a_readback_error_after_set_removes_the_reserved_record(tmp_path):
+    backend = FakeSecretBackend()
+    credentials, _ = make_store(tmp_path, backend)
+    name = record_name(credentials, 1)
+    backend.fail_get.add(name)
+
+    with pytest.raises(BackendOperationError, match="read failed"):
+        credentials.replace("india_prod", "key", "secret")
+
+    metadata = credentials.metadata("india_prod")
+    assert metadata.revision is None
+    assert metadata.pending_revisions == ()
+    assert backend.values == {}
+
+
+def test_failed_readback_cleanup_remains_retryable_after_restart(
+    tmp_path,
+    monkeypatch,
+):
+    backend = FakeSecretBackend()
+    credentials, _ = make_store(tmp_path, backend)
+    name = record_name(credentials, 1)
+    backend.fail_get.add(name)
+    backend.fail_delete.add(name)
+
+    with pytest.raises(CredentialStoreError, match="cleanup remains pending"):
+        credentials.replace("india_prod", "key", "secret")
+
+    pending = credentials.metadata("india_prod")
+    assert pending.revision is None
+    assert pending.pending_revisions == (1,)
+    assert set(backend.values) == {name}
+
+    monkeypatch.setattr(auth_store, "SystemKeyringBackend", lambda: backend)
+    restarted = CredentialStore.open(tmp_path / "credentials.json")
+
+    assert restarted.metadata("india_prod").pending_revisions == ()
+    assert restarted.get("india_prod") is None
+    assert backend.values == {}
+
+
+def test_process_death_during_rotation_readback_preserves_the_previous_record(
+    tmp_path,
+    monkeypatch,
+):
+    backend = FakeSecretBackend()
+    credentials, _ = make_store(tmp_path, backend)
+    first = credentials.replace("india_prod", "old-key", "old-secret")
+    old_name = record_name(credentials, 1)
+    name = record_name(credentials, 2)
+    backend.crash_get.add(name)
+
+    with pytest.raises(SimulatedProcessDeath):
+        credentials.replace(
+            "india_prod",
+            "new-key",
+            "new-secret",
+            expected_revision=first.revision,
+        )
+
+    pending = credentials.metadata("india_prod")
+    assert (pending.revision, pending.generation) == (1, 1)
+    assert pending.pending_revisions == (2,)
+    assert set(backend.values) == {old_name, name}
+
+    monkeypatch.setattr(auth_store, "SystemKeyringBackend", lambda: backend)
+    restarted = CredentialStore.open(tmp_path / "credentials.json")
+
+    assert restarted.metadata("india_prod").pending_revisions == ()
+    assert restarted.get("india_prod") == first
+    assert set(backend.values) == {old_name}
 
 
 def test_old_record_delete_failure_rolls_the_rotation_back(tmp_path):
@@ -885,455 +591,6 @@ def test_a_partial_process_pair_fails_without_falling_through_to_the_store(tmp_p
 
     with pytest.raises(IncompleteCredentialError, match="must supply"):
         credentials.resolve("india_prod", {"DELTA_API_KEY": "only-a-key"})
-
-
-def test_no_approved_system_backend_uses_process_local_memory(monkeypatch, tmp_path):
-    def unavailable():
-        raise BackendUnavailableError("no secure backend")
-
-    monkeypatch.setattr(auth_store, "SystemKeyringBackend", unavailable)
-    credentials = CredentialStore.open(tmp_path / "credentials.json")
-
-    saved = credentials.replace("india_prod", "key", "secret")
-
-    assert credentials.persistent is False
-    assert credentials.source is CredentialSource.MEMORY
-    assert credentials.fallback_reason == "no secure backend"
-    assert saved.session_only is True
-    assert credentials.get("india_prod") == saved
-    assert not (tmp_path / "credentials.json").exists()
-
-
-def test_keyring_discovery_failure_uses_process_local_memory(monkeypatch, tmp_path):
-    def fail_discovery():
-        raise RuntimeError("backend discovery failed")
-
-    monkeypatch.setattr(auth_store.keyring, "get_keyring", fail_discovery)
-
-    credentials = CredentialStore.open(tmp_path / "credentials.json")
-
-    assert credentials.source is CredentialSource.MEMORY
-    assert "backend discovery failed" in credentials.fallback_reason
-    assert not (tmp_path / "credentials.json").exists()
-
-
-def test_memory_fallback_never_removes_the_only_persistent_copy(tmp_path):
-    credentials = CredentialStore(
-        auth_store.MemorySecretBackend(),
-        MemoryMetadata(),
-        CredentialSource.MEMORY,
-    )
-    config_path = tmp_path / "config.env"
-    original = "DELTA_API_KEY=key\nDELTA_API_SECRET=secret\n"
-    config_path.write_text(original)
-
-    result = credentials.migrate(config_path)
-
-    assert result.status is MigrationStatus.UNAVAILABLE
-    assert config_path.read_text() == original
-    assert credentials.get("india_prod") is None
-
-
-class NullKeyring(KeyringBackend):
-    priority = 0
-
-    def get_password(self, service, username):
-        return None
-
-    def set_password(self, service, username, password):
-        return None
-
-    def delete_password(self, service, username):
-        return None
-
-
-class PlaintextKeyring(NullKeyring):
-    __module__ = "keyrings.alt.file"
-    priority = 1
-
-
-@pytest.mark.parametrize("backend", [NullKeyring(), PlaintextKeyring()])
-def test_null_and_plaintext_keyrings_are_rejected(backend):
-    with pytest.raises(BackendUnavailableError, match="not an approved"):
-        SystemKeyringBackend(backend)
-
-
-def test_successful_migration_removes_only_credential_lines(tmp_path):
-    credentials, _ = make_store(tmp_path)
-    config_path = tmp_path / "config.env"
-    original = (
-        "# keep this comment\n"
-        'export DELTA_API_KEY = "legacy-key"\n'
-        "DELTA_API_SECRET='legacy-secret'\n"
-        "DELTA_MCP_ENV=india_testnet\n"
-        "DELTA_MCP_MODE=trade\n"
-        "UNRELATED=value\n"
-    )
-    config_path.write_text(original)
-
-    result = credentials.migrate(config_path)
-
-    assert result.status is MigrationStatus.MIGRATED
-    assert result.environment == "india_testnet"
-    assert result.credential is not None
-    assert result.credential.state is CredentialState.UNVERIFIED
-    assert (result.credential.api_key, result.credential.api_secret) == (
-        "legacy-key",
-        "legacy-secret",
-    )
-    assert config_path.read_text() == (
-        "# keep this comment\n"
-        "DELTA_MCP_ENV=india_testnet\n"
-        "DELTA_MCP_MODE=trade\n"
-        "UNRELATED=value\n"
-    )
-
-
-def test_migration_rejects_a_symlink_without_leaving_secrets_in_its_target(
-    tmp_path,
-) -> None:
-    credentials, backend = make_store(tmp_path)
-    target = tmp_path / "real-config.env"
-    target.write_text(
-        "DELTA_API_KEY=legacy-key\n"
-        "DELTA_API_SECRET=legacy-secret\n"
-        "DELTA_MCP_ENV=india_prod\n"
-    )
-    config_path = tmp_path / "config.env"
-    config_path.symlink_to(target)
-
-    result = credentials.migrate(config_path)
-
-    assert result.status is MigrationStatus.UNAVAILABLE
-    assert "DELTA_API_KEY=legacy-key" in target.read_text()
-    assert "DELTA_API_SECRET=legacy-secret" in target.read_text()
-    assert backend.values == {}
-
-
-def test_migration_removes_a_multiline_quoted_secret_as_one_setting(tmp_path):
-    credentials, _ = make_store(tmp_path)
-    config_path = tmp_path / "config.env"
-    config_path.write_text(
-        "DELTA_API_KEY=key\n"
-        "DELTA_API_SECRET='first line\nsecond line'\n"
-        "UNRELATED=value\n"
-    )
-
-    result = credentials.migrate(config_path)
-
-    assert result.status is MigrationStatus.MIGRATED
-    assert config_path.read_text() == "UNRELATED=value\n"
-    saved = credentials.get("india_prod")
-    assert saved is not None
-    assert saved.api_secret == "first line\nsecond line"
-
-
-def test_migration_of_an_incomplete_pair_leaves_the_file_unchanged(tmp_path):
-    credentials, _ = make_store(tmp_path)
-    config_path = tmp_path / "config.env"
-    original = "DELTA_API_KEY=only-a-key\nDELTA_MCP_MODE=trade\n"
-    config_path.write_text(original)
-
-    result = credentials.migrate(config_path)
-
-    assert result.status is MigrationStatus.INCOMPLETE
-    assert config_path.read_text() == original
-    assert credentials.get("india_prod") is None
-
-
-def test_migration_readback_failure_leaves_the_file_unchanged(tmp_path):
-    backend = FakeSecretBackend()
-    credentials, _ = make_store(tmp_path, backend)
-    backend.mismatch.add(record_name(credentials, 1))
-    config_path = tmp_path / "config.env"
-    original = "DELTA_API_KEY=key\nDELTA_API_SECRET=secret\nDELTA_MCP_MODE=trade\n"
-    config_path.write_text(original)
-
-    with pytest.raises(BackendOperationError, match="did not return"):
-        credentials.migrate(config_path)
-
-    assert config_path.read_text() == original
-    assert credentials.get("india_prod") is None
-
-
-def test_migration_publish_failure_rolls_back_the_new_record(tmp_path, monkeypatch):
-    credentials, backend = make_store(tmp_path)
-    config_path = tmp_path / "config.env"
-    original = "DELTA_API_KEY=key\nDELTA_API_SECRET=secret\nDELTA_MCP_MODE=trade\n"
-    config_path.write_text(original)
-    real_replace = os.replace
-
-    def fail_config_publish(source, target):
-        if target == config_path:
-            raise OSError("read-only config")
-        return real_replace(source, target)
-
-    monkeypatch.setattr(auth_store.os, "replace", fail_config_publish)
-
-    with pytest.raises(auth_store.MigrationError, match="read-only config"):
-        credentials.migrate(config_path)
-
-    assert config_path.read_text() == original
-    assert credentials.get("india_prod") is None
-    assert backend.values == {}
-
-
-def test_migration_publish_failure_restores_the_prior_active_record(
-    tmp_path,
-    monkeypatch,
-):
-    credentials, backend = make_store(tmp_path)
-    first = credentials.replace("india_prod", "same-key", "same-secret")
-    config_path = tmp_path / "config.env"
-    original = (
-        "DELTA_API_KEY=same-key\nDELTA_API_SECRET=same-secret\nDELTA_MCP_MODE=trade\n"
-    )
-    config_path.write_text(original)
-    real_replace = os.replace
-
-    def fail_config_publish(source, target):
-        if target == config_path:
-            raise OSError("read-only config")
-        return real_replace(source, target)
-
-    monkeypatch.setattr(auth_store.os, "replace", fail_config_publish)
-
-    with pytest.raises(auth_store.MigrationError, match="read-only config"):
-        credentials.migrate(config_path)
-
-    assert config_path.read_text() == original
-    assert credentials.get("india_prod") == first
-    assert set(backend.values) == {record_name(credentials, 1)}
-
-
-def test_config_directory_sync_failure_does_not_rollback_published_migration(
-    tmp_path,
-    monkeypatch,
-):
-    backend = FakeSecretBackend()
-    credentials = CredentialStore(
-        backend,
-        FileMetadata(tmp_path / "metadata" / "credentials.json"),
-        CredentialSource.OS_STORE,
-    )
-    config_path = tmp_path / "config.env"
-    config_path.write_text(
-        "DELTA_API_KEY=key\nDELTA_API_SECRET=secret\nDELTA_MCP_MODE=trade\n"
-    )
-    real_sync = auth_store._sync_directory
-    config_sync_failed = False
-
-    def fail_config_directory_sync(path):
-        nonlocal config_sync_failed
-        if path == config_path.parent:
-            config_sync_failed = True
-            raise OSError("config directory sync failed")
-        real_sync(path)
-
-    monkeypatch.setattr(auth_store, "_sync_directory", fail_config_directory_sync)
-
-    result = credentials.migrate(config_path)
-
-    assert result.status is MigrationStatus.MIGRATED
-    assert config_sync_failed is True
-    assert config_path.read_text() == "DELTA_MCP_MODE=trade\n"
-    assert credentials.get("india_prod") == result.credential
-    assert set(backend.values) == {record_name(credentials, 1)}
-
-
-def test_migration_of_the_active_pair_advances_revision_and_generation(tmp_path):
-    credentials, backend = make_store(tmp_path)
-    first = credentials.replace(
-        "india_prod",
-        "same-key",
-        "same-secret",
-        state=CredentialState.VERIFIED,
-        account_id="account-123",
-    )
-    config_path = tmp_path / "config.env"
-    config_path.write_text(
-        "DELTA_API_KEY=same-key\nDELTA_API_SECRET=same-secret\nDELTA_MCP_MODE=trade\n"
-    )
-
-    result = credentials.migrate(config_path)
-
-    assert result.status is MigrationStatus.MIGRATED
-    assert result.credential is not None
-    assert (result.credential.revision, result.credential.generation) == (2, 2)
-    assert result.credential.state is CredentialState.VERIFIED
-    assert result.credential.account_id == "account-123"
-    assert first.revision == 1
-    assert config_path.read_text() == "DELTA_MCP_MODE=trade\n"
-    assert set(backend.values) == {record_name(credentials, 2)}
-
-
-@pytest.mark.parametrize("failure_mode", ["before_delete", "after_delete"])
-def test_migration_stays_complete_when_inactive_record_retirement_fails(
-    tmp_path,
-    caplog,
-    failure_mode,
-):
-    credentials, backend = make_store(tmp_path)
-    first = credentials.replace("india_prod", "same-key", "same-secret")
-    old_name = record_name(credentials, 1)
-    if failure_mode == "before_delete":
-        backend.fail_delete.add(old_name)
-        expected_records = {old_name, record_name(credentials, 2)}
-    else:
-        backend.fail_delete_after.add(old_name)
-        expected_records = {record_name(credentials, 2)}
-    config_path = tmp_path / "config.env"
-    config_path.write_text(
-        "DELTA_API_KEY=same-key\nDELTA_API_SECRET=same-secret\nDELTA_MCP_MODE=trade\n"
-    )
-
-    result = credentials.migrate(config_path)
-
-    assert result.status is MigrationStatus.MIGRATED
-    assert result.credential is not None
-    assert (result.credential.revision, result.credential.generation) == (2, 2)
-    assert config_path.read_text() == "DELTA_MCP_MODE=trade\n"
-    assert set(backend.values) == expected_records
-    assert credentials.metadata("india_prod").pending_revisions == (1,)
-    assert credentials.get("india_prod") == result.credential
-    assert credentials.metadata("india_prod").pending_revisions == ()
-    assert set(backend.values) == {record_name(credentials, 2)}
-    assert first.revision == 1
-    assert "could not retire inactive credential revision 1" in caplog.text
-    assert "same-key" not in caplog.text
-    assert "same-secret" not in caplog.text
-
-
-@pytest.mark.parametrize("operation", ["delete", "rotate"])
-def test_migration_serializes_concurrent_credential_changes(
-    tmp_path,
-    monkeypatch,
-    operation,
-):
-    credentials, backend = make_store(tmp_path)
-    first = credentials.replace("india_prod", "same-key", "same-secret")
-    other_process_view = CredentialStore(
-        backend,
-        FileMetadata(tmp_path / "credentials.json"),
-        CredentialSource.OS_STORE,
-    )
-    config_path = tmp_path / "config.env"
-    config_path.write_text("DELTA_API_KEY=same-key\nDELTA_API_SECRET=same-secret\n")
-    config_publish_started = threading.Event()
-    release_config_publish = threading.Event()
-    credential_change_started = threading.Event()
-    migration_results = []
-    outcomes: list[str] = []
-    errors: list[Exception] = []
-    real_replace = os.replace
-
-    def pause_config_publish(source, target):
-        if target == config_path:
-            config_publish_started.set()
-            if not release_config_publish.wait(5):
-                raise OSError("timed out waiting to publish config")
-        return real_replace(source, target)
-
-    def run_migration():
-        try:
-            migration_results.append(credentials.migrate(config_path))
-        except Exception as exc:
-            errors.append(exc)
-
-    def change_credential():
-        credential_change_started.set()
-        try:
-            if operation == "delete":
-                other_process_view.delete(
-                    "india_prod",
-                    expected_revision=first.revision,
-                )
-            else:
-                other_process_view.replace(
-                    "india_prod",
-                    "rotated-key",
-                    "rotated-secret",
-                    expected_revision=first.revision,
-                )
-            outcomes.append("changed")
-        except CredentialConflictError:
-            outcomes.append("conflict")
-        except Exception as exc:
-            errors.append(exc)
-
-    monkeypatch.setattr(auth_store.os, "replace", pause_config_publish)
-    migration_thread = threading.Thread(target=run_migration)
-    migration_thread.start()
-    assert config_publish_started.wait(2)
-
-    change_thread = threading.Thread(target=change_credential)
-    change_thread.start()
-    assert credential_change_started.wait(2)
-    change_thread.join(0.1)
-    try:
-        assert change_thread.is_alive()
-    finally:
-        release_config_publish.set()
-
-    migration_thread.join(2)
-    change_thread.join(2)
-    assert not migration_thread.is_alive()
-    assert not change_thread.is_alive()
-    assert errors == []
-    assert outcomes == ["conflict"]
-    assert len(migration_results) == 1
-    migrated = migration_results[0]
-    assert migrated.status is MigrationStatus.MIGRATED
-    assert migrated.credential is not None
-    assert (migrated.credential.revision, migrated.credential.generation) == (2, 2)
-    assert credentials.get("india_prod") == migrated.credential
-    assert set(backend.values) == {record_name(credentials, 2)}
-
-
-def test_migration_never_overwrites_a_different_active_credential(tmp_path):
-    credentials, _ = make_store(tmp_path)
-    active = credentials.replace("india_prod", "current-key", "current-secret")
-    config_path = tmp_path / "config.env"
-    original = "DELTA_API_KEY=old-key\nDELTA_API_SECRET=old-secret\n"
-    config_path.write_text(original)
-
-    result = credentials.migrate(config_path)
-
-    assert result.status is MigrationStatus.CONFLICT
-    assert result.credential == active
-    assert config_path.read_text() == original
-    assert credentials.get("india_prod") == active
-
-
-def test_migration_observes_a_concurrent_create_before_its_transaction(
-    tmp_path,
-    monkeypatch,
-):
-    credentials, backend = make_store(tmp_path)
-    other_process_view = CredentialStore(
-        backend,
-        FileMetadata(tmp_path / "credentials.json"),
-        CredentialSource.OS_STORE,
-    )
-    config_path = tmp_path / "config.env"
-    original = "DELTA_API_KEY=file-key\nDELTA_API_SECRET=file-secret\n"
-    config_path.write_text(original)
-    real_stage = auth_store._stage_replacement
-
-    def stage_after_browser_write(path, body):
-        staged = real_stage(path, body)
-        other_process_view.replace("india_prod", "browser-key", "browser-secret")
-        return staged
-
-    monkeypatch.setattr(auth_store, "_stage_replacement", stage_after_browser_write)
-
-    result = credentials.migrate(config_path)
-
-    assert result.status is MigrationStatus.CONFLICT
-    assert result.credential is not None
-    assert result.credential.api_key == "browser-key"
-    assert config_path.read_text() == original
-    assert set(backend.values) == {record_name(credentials, 1)}
 
 
 @pytest.mark.parametrize("environment", ["india_devnet", "other", "../india_prod"])
